+++ RSSReaderTests/RSSReaderTests.swift
@@ -7,733 +7,6 @@ import Testing
 @MainActor
 struct RSSReaderTests {
     @Test
-    func singleFeedRefreshFetchedPersistsArticlesMetadataAndFetchState() async throws {
-        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                steps: [
-                    .response(
-                        statusCode: 200,
-                        headers: [
-                            "Content-Type": "application/rss+xml; charset=utf-8",
-                            "ETag": "\"etag-new\"",
-                            "Last-Modified": "Tue, 02 Jan 2024 12:00:00 GMT"
-                        ],
-                        body: Self.validRSSFeedXML(
-                            channelTitle: "Updated Feed Title",
-                            channelLink: "https://example.com/",
-                            language: "en",
-                            itemTitle: "Article One",
-                            itemLink: "https://example.com/articles/1",
-                            itemGUID: "article-1",
-                            itemDescription: "Readable summary",
-                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                        )
-                    )
-                ]
-            )
-        )
-
-        let feed = Feed(
-            url: "https://example.com/feed.xml",
-            title: "Old Feed Title",
-            lastSuccessfulFetchAt: oldSuccessAt,
-            lastETag: "\"etag-old\"",
-            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT",
-            lastSyncError: "Previous error"
-        )
-        try harness.feedRepository.insert(feed)
-
-        let result = await harness.service.refresh(feedID: feed.id)
-
-        #expect(result.status == .fetched)
-        #expect(result.processedEntryCount == 1)
-        #expect(result.upsertedEntryCount == 1)
-        #expect(result.rejectedEntryCount == 0)
-        #expect(result.errorDescription == nil)
-
-        let fetchedFeed = try harness.fetchFeed(id: feed.id)
-        let refreshedFeed = try #require(fetchedFeed)
-        #expect(refreshedFeed.title == "Updated Feed Title")
-        #expect(refreshedFeed.siteURL == "https://example.com/")
-        #expect(refreshedFeed.language == "en")
-        #expect(refreshedFeed.kind == .rss)
-        #expect(refreshedFeed.lastFetchedAt != nil)
-        #expect(refreshedFeed.lastSuccessfulFetchAt != nil)
-        #expect(refreshedFeed.lastSuccessfulFetchAt != oldSuccessAt)
-        #expect(refreshedFeed.lastETag == "\"etag-new\"")
-        #expect(refreshedFeed.lastModifiedHeader == "Tue, 02 Jan 2024 12:00:00 GMT")
-        #expect(refreshedFeed.lastSyncError == nil)
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
-        #expect(articles.count == 1)
-        #expect(articles.first?.title == "Article One")
-        #expect(articles.first?.isDeletedAtSource == false)
-
-        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
-        let latestLog = try #require(fetchedLog)
-        #expect(latestLog.status == "fetched")
-        #expect(latestLog.httpCode == 200)
-
-        let requests = await harness.httpClient.recordedRequests()
-        #expect(requests.count == 1)
-        #expect(requests.first?.headers["If-None-Match"] == "\"etag-old\"")
-        #expect(requests.first?.headers["If-Modified-Since"] == "Mon, 01 Jan 2024 12:00:00 GMT")
-    }
-
-    @Test
-    func singleFeedRefreshNotModifiedUpdatesFetchStateWithoutParsingPipeline() async throws {
-        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                steps: [
-                    .response(
-                        statusCode: 304,
-                        headers: [
-                            "ETag": "\"etag-304\"",
-                            "Last-Modified": "Wed, 03 Jan 2024 12:00:00 GMT"
-                        ],
-                        body: ""
-                    )
-                ]
-            )
-        )
-
-        let feed = Feed(
-            url: "https://example.com/feed.xml",
-            title: "Stable Feed Title",
-            lastSuccessfulFetchAt: oldSuccessAt,
-            lastETag: "\"etag-old\"",
-            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT",
-            lastSyncError: "Transient error"
-        )
-        try harness.feedRepository.insert(feed)
-
-        let result = await harness.service.refresh(feedID: feed.id)
-
-        #expect(result.status == .notModified)
-        #expect(result.processedEntryCount == 0)
-        #expect(result.upsertedEntryCount == 0)
-        #expect(result.rejectedEntryCount == 0)
-        #expect(result.errorDescription == nil)
-
-        let fetchedFeed = try harness.fetchFeed(id: feed.id)
-        let refreshedFeed = try #require(fetchedFeed)
-        #expect(refreshedFeed.title == "Stable Feed Title")
-        #expect(refreshedFeed.lastFetchedAt != nil)
-        #expect(refreshedFeed.lastSuccessfulFetchAt == oldSuccessAt)
-        #expect(refreshedFeed.lastETag == "\"etag-304\"")
-        #expect(refreshedFeed.lastModifiedHeader == "Wed, 03 Jan 2024 12:00:00 GMT")
-        #expect(refreshedFeed.lastSyncError == nil)
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
-        #expect(articles.isEmpty)
-
-        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
-        let latestLog = try #require(fetchedLog)
-        #expect(latestLog.status == "not_modified")
-        #expect(latestLog.httpCode == 304)
-
-        let requests = await harness.httpClient.recordedRequests()
-        #expect(requests.count == 1)
-        #expect(requests.first?.headers["If-None-Match"] == "\"etag-old\"")
-        #expect(requests.first?.headers["If-Modified-Since"] == "Mon, 01 Jan 2024 12:00:00 GMT")
-    }
-
-    @Test
-    func singleFeedRefreshFailedPersistsErrorWithoutWritingArticles() async throws {
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                steps: [
-                    .response(
-                        statusCode: 500,
-                        headers: [
-                            "Content-Type": "application/rss+xml; charset=utf-8"
-                        ],
-                        body: ""
-                    )
-                ]
-            )
-        )
-
-        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
-        let feed = Feed(
-            url: "https://example.com/feed.xml",
-            title: "Failing Feed",
-            lastSuccessfulFetchAt: oldSuccessAt,
-            lastETag: "\"etag-old\"",
-            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT"
-        )
-        try harness.feedRepository.insert(feed)
-
-        let result = await harness.service.refresh(feedID: feed.id)
-
-        #expect(result.status == .failed)
-        #expect(result.processedEntryCount == 0)
-        #expect(result.upsertedEntryCount == 0)
-        #expect(result.rejectedEntryCount == 0)
-        #expect(result.errorDescription?.contains("invalidStatusCode") == true)
-
-        let fetchedFeed = try harness.fetchFeed(id: feed.id)
-        let refreshedFeed = try #require(fetchedFeed)
-        #expect(refreshedFeed.title == "Failing Feed")
-        #expect(refreshedFeed.lastFetchedAt != nil)
-        #expect(refreshedFeed.lastSuccessfulFetchAt == oldSuccessAt)
-        #expect(refreshedFeed.lastSyncError?.contains("invalidStatusCode") == true)
-        #expect(refreshedFeed.lastETag == "\"etag-old\"")
-        #expect(refreshedFeed.lastModifiedHeader == "Mon, 01 Jan 2024 12:00:00 GMT")
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
-        #expect(articles.isEmpty)
-
-        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
-        let latestLog = try #require(fetchedLog)
-        #expect(latestLog.status == "failed")
-        #expect(latestLog.httpCode == 500)
-        #expect(latestLog.message?.contains("invalidStatusCode") == true)
-    }
-
-    @Test
-    func singleFeedRefreshCancelledReturnsCancelledWithoutPersistingFailureState() async throws {
-        let oldSuccessAt = Date(timeIntervalSince1970: 1_700_000_000)
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                steps: [
-                    .cancelled
-                ]
-            )
-        )
-
-        let feed = Feed(
-            url: "https://example.com/feed.xml",
-            title: "Cancellable Feed",
-            lastSuccessfulFetchAt: oldSuccessAt,
-            lastETag: "\"etag-old\"",
-            lastModifiedHeader: "Mon, 01 Jan 2024 12:00:00 GMT",
-            lastSyncError: "Previous error"
-        )
-        try harness.feedRepository.insert(feed)
-
-        let result = await harness.service.refresh(feedID: feed.id)
-
-        #expect(result.status == .cancelled)
-        #expect(result.processedEntryCount == 0)
-        #expect(result.upsertedEntryCount == 0)
-        #expect(result.rejectedEntryCount == 0)
-        #expect(result.errorDescription == "Refresh cancelled")
-
-        let fetchedFeed = try harness.fetchFeed(id: feed.id)
-        let refreshedFeed = try #require(fetchedFeed)
-        #expect(refreshedFeed.title == "Cancellable Feed")
-        #expect(refreshedFeed.lastFetchedAt != nil)
-        #expect(refreshedFeed.lastSuccessfulFetchAt == oldSuccessAt)
-        #expect(refreshedFeed.lastSyncError == "Previous error")
-        #expect(refreshedFeed.lastETag == "\"etag-old\"")
-        #expect(refreshedFeed.lastModifiedHeader == "Mon, 01 Jan 2024 12:00:00 GMT")
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
-        #expect(articles.isEmpty)
-
-        let fetchedLog = try harness.feedFetchLogRepository.fetchLatestLog(feedID: feed.id)
-        let latestLog = try #require(fetchedLog)
-        #expect(latestLog.status == "cancelled")
-        #expect(latestLog.httpCode == nil)
-        #expect(latestLog.message?.contains("Refresh cancelled") == true)
-    }
-
-    @Test
-    func batchRefreshAggregatesPartialSuccessAndIndividualFailures() async throws {
-        let feed1URL = "https://example.com/feed-1.xml"
-        let feed2URL = "https://example.com/feed-2.xml"
-        let feed3URL = "https://example.com/feed-3.xml"
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                responsesByURL: [
-                    feed1URL: .response(
-                        statusCode: 200,
-                        headers: [
-                            "Content-Type": "application/rss+xml; charset=utf-8",
-                            "ETag": "\"etag-feed-1\""
-                        ],
-                        body: Self.validRSSFeedXML(
-                            channelTitle: "Feed One",
-                            channelLink: "https://example.com/one/",
-                            language: "en",
-                            itemTitle: "Batch Article One",
-                            itemLink: "https://example.com/one/articles/1",
-                            itemGUID: "batch-article-1",
-                            itemDescription: "Readable summary one",
-                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                        )
-                    ),
-                    feed2URL: .response(
-                        statusCode: 304,
-                        headers: [
-                            "ETag": "\"etag-feed-2\"",
-                            "Last-Modified": "Wed, 03 Jan 2024 12:00:00 GMT"
-                        ],
-                        body: ""
-                    ),
-                    feed3URL: .response(
-                        statusCode: 500,
-                        headers: [
-                            "Content-Type": "application/rss+xml; charset=utf-8"
-                        ],
-                        body: ""
-                    )
-                ]
-            )
-        )
-
-        let feeds = try harness.insertFeeds(urls: [feed1URL, feed2URL, feed3URL])
-
-        let result = await harness.service.refreshFeeds(feeds.map(\.id))
-
-        #expect(result.summary.totalFeedCount == 3)
-        #expect(result.summary.fetchedCount == 1)
-        #expect(result.summary.notModifiedCount == 1)
-        #expect(result.summary.failedCount == 1)
-        #expect(result.summary.cancelledCount == 0)
-        #expect(result.errors.count == 1)
-        #expect(result.failedResults.count == 1)
-        #expect(result.results.map(\.status) == [.fetched, .notModified, .failed])
-        #expect(result.errors.first?.feedID == feeds[2].id)
-        #expect(result.errors.first?.message.contains("invalidStatusCode") == true)
-
-        let feed1Articles = try harness.articleRepository.fetchArticles(feedID: feeds[0].id)
-        #expect(feed1Articles.count == 1)
-
-        let fetchedFeed2 = try harness.fetchFeed(id: feeds[1].id)
-        let feed2State = try #require(fetchedFeed2)
-        #expect(feed2State.lastETag == "\"etag-feed-2\"")
-
-        let fetchedFeed3 = try harness.fetchFeed(id: feeds[2].id)
-        let feed3State = try #require(fetchedFeed3)
-        #expect(feed3State.lastSyncError?.contains("invalidStatusCode") == true)
-    }
-
-    @Test
-    func batchRefreshRespectsDefaultConcurrencyLimit() async throws {
-        let urls = (1...4).map { "https://example.com/concurrency-\($0).xml" }
-        let responses = Dictionary(uniqueKeysWithValues: urls.enumerated().map { index, url in
-            (
-                url,
-                ScriptedHTTPClient.Step.delayedResponse(
-                    statusCode: 200,
-                    headers: [
-                        "Content-Type": "application/rss+xml; charset=utf-8"
-                    ],
-                    body: Self.validRSSFeedXML(
-                        channelTitle: "Concurrency Feed \(index + 1)",
-                        channelLink: "https://example.com/\(index + 1)/",
-                        language: "en",
-                        itemTitle: "Concurrency Article \(index + 1)",
-                        itemLink: "https://example.com/\(index + 1)/articles/1",
-                        itemGUID: "concurrency-\(index + 1)",
-                        itemDescription: "Readable summary \(index + 1)",
-                        pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                    ),
-                    delayNanoseconds: 200_000_000
-                )
-            )
-        })
-
-        let client = ScriptedHTTPClient(responsesByURL: responses)
-        let harness = try TestHarness.make(httpClient: client)
-        let feeds = try harness.insertFeeds(urls: urls)
-
-        let result = await harness.service.refreshFeeds(feeds.map(\.id))
-
-        #expect(result.summary.totalFeedCount == 4)
-        #expect(result.summary.fetchedCount == 4)
-        #expect(result.summary.failedCount == 0)
-        #expect(result.summary.cancelledCount == 0)
-
-        let maxConcurrentExecutions = await client.maxConcurrentExecutions()
-        #expect(maxConcurrentExecutions <= 3)
-
-        let requests = await client.recordedRequests()
-        #expect(requests.count == 4)
-    }
-
-    @Test
-    func batchRefreshCancellationReturnsPartialCancelledResults() async throws {
-        let urls = (1...4).map { "https://example.com/cancel-\($0).xml" }
-        let responses = Dictionary(uniqueKeysWithValues: urls.enumerated().map { index, url in
-            (
-                url,
-                ScriptedHTTPClient.Step.delayedResponse(
-                    statusCode: 200,
-                    headers: [
-                        "Content-Type": "application/rss+xml; charset=utf-8"
-                    ],
-                    body: Self.validRSSFeedXML(
-                        channelTitle: "Cancel Feed \(index + 1)",
-                        channelLink: "https://example.com/cancel/\(index + 1)/",
-                        language: "en",
-                        itemTitle: "Cancel Article \(index + 1)",
-                        itemLink: "https://example.com/cancel/\(index + 1)/articles/1",
-                        itemGUID: "cancel-\(index + 1)",
-                        itemDescription: "Readable summary \(index + 1)",
-                        pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                    ),
-                    delayNanoseconds: 500_000_000
-                )
-            )
-        })
-
-        let client = ScriptedHTTPClient(responsesByURL: responses)
-        let harness = try TestHarness.make(httpClient: client)
-        let feeds = try harness.insertFeeds(urls: urls)
-
-        let task = Task { @MainActor in
-            await harness.service.refreshFeeds(feeds.map(\.id))
-        }
-
-        try await Task.sleep(nanoseconds: 50_000_000)
-        task.cancel()
-
-        let result = await task.value
-
-        #expect(result.summary.totalFeedCount > 0)
-        #expect(result.summary.totalFeedCount < 4)
-        #expect(result.summary.fetchedCount + result.summary.cancelledCount == result.summary.totalFeedCount)
-        #expect(result.summary.failedCount == 0)
-        #expect(result.summary.notModifiedCount == 0)
-
-        let requests = await client.recordedRequests()
-        #expect(requests.count <= 3)
-    }
-
-    @Test
-    func concurrentRefreshOfSameFeedSharesInFlightTaskAndAvoidsDuplicateSideEffects() async throws {
-        let feedURL = "https://example.com/concurrent-feed.xml"
-        let client = ScriptedHTTPClient(
-            responsesByURL: [
-                feedURL: .delayedResponse(
-                    statusCode: 200,
-                    headers: [
-                        "Content-Type": "application/rss+xml; charset=utf-8",
-                        "ETag": "\"etag-concurrent\""
-                    ],
-                    body: Self.validRSSFeedXML(
-                        channelTitle: "Concurrent Feed",
-                        channelLink: "https://example.com/concurrent/",
-                        language: "en",
-                        itemTitle: "Concurrent Article",
-                        itemLink: "https://example.com/concurrent/articles/1",
-                        itemGUID: "concurrent-article-1",
-                        itemDescription: "Readable concurrent summary",
-                        pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                    ),
-                    delayNanoseconds: 200_000_000
-                )
-            ]
-        )
-        let harness = try TestHarness.make(httpClient: client)
-        let feed = try harness.insertFeeds(urls: [feedURL]).first
-        let requiredFeed = try #require(feed)
-
-        let firstTask = Task { @MainActor in
-            await harness.service.refresh(feedID: requiredFeed.id)
-        }
-        let secondTask = Task { @MainActor in
-            await harness.service.refresh(feedID: requiredFeed.id)
-        }
-
-        let firstResult = await firstTask.value
-        let secondResult = await secondTask.value
-
-        #expect(firstResult.status == .fetched)
-        #expect(secondResult.status == .fetched)
-        #expect(firstResult.upsertedEntryCount == 1)
-        #expect(secondResult.upsertedEntryCount == 1)
-        #expect(firstResult.finishedAt == secondResult.finishedAt)
-
-        let requests = await client.recordedRequests()
-        #expect(requests.count == 1)
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: requiredFeed.id)
-        #expect(articles.count == 1)
-
-        let logs = try harness.feedFetchLogRepository.fetchLogs(feedID: requiredFeed.id, limit: nil)
-        #expect(logs.count == 1)
-        #expect(logs.first?.status == "fetched")
-    }
-
-    @Test
-    func refreshUpdatesFeedMetadataAndMarksMissingArticlesAsDeletedAtSource() async throws {
-        let feedURL = "https://example.com/reconcile-feed.xml"
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                responsesByURL: [
-                    feedURL: .response(
-                        statusCode: 200,
-                        headers: [
-                            "Content-Type": "application/rss+xml; charset=utf-8"
-                        ],
-                        body: Self.validRSSFeedXML(
-                            channelTitle: "Reconciled Feed Title",
-                            channelLink: "https://example.com/reconciled/",
-                            language: "fr",
-                            itemTitle: "Current Article",
-                            itemLink: "https://example.com/reconciled/articles/current",
-                            itemGUID: "current-article",
-                            itemDescription: "Current article summary",
-                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                        )
-                    )
-                ]
-            )
-        )
-
-        let feed = try #require(try harness.insertFeeds(urls: [feedURL]).first)
-        feed.title = "Stale Feed Title"
-        feed.siteURL = "https://example.com/old/"
-        feed.language = "en"
-        feed.kind = .unknown
-        try harness.saveModelContext()
-
-        _ = try harness.insertArticle(
-            feed: feed,
-            externalID: "obsolete-article",
-            guid: "obsolete-article",
-            url: "https://example.com/reconciled/articles/obsolete",
-            title: "Obsolete Article"
-        )
-
-        let result = await harness.service.refresh(feedID: feed.id)
-
-        #expect(result.status == .fetched)
-        #expect(result.upsertedEntryCount == 1)
-
-        let refreshedFeed = try #require(try harness.fetchFeed(id: feed.id))
-        #expect(refreshedFeed.title == "Reconciled Feed Title")
-        #expect(refreshedFeed.siteURL == "https://example.com/reconciled/")
-        #expect(refreshedFeed.language == "fr")
-        #expect(refreshedFeed.kind == .rss)
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
-        #expect(articles.count == 2)
-
-        let obsoleteArticle = try #require(articles.first { $0.externalID == "obsolete-article" })
-        #expect(obsoleteArticle.isDeletedAtSource == true)
-
-        let currentArticle = try #require(articles.first { $0.guid == "current-article" })
-        #expect(currentArticle.isDeletedAtSource == false)
-        #expect(currentArticle.title == "Current Article")
-    }
-
-    @Test
-    func refreshReactivatesArticleWhenItReappearsInFeedPayload() async throws {
-        let feedURL = "https://example.com/reappearing-feed.xml"
-        let harness = try TestHarness.make(
-            httpClient: ScriptedHTTPClient(
-                responsesByURL: [
-                    feedURL: .response(
-                        statusCode: 200,
-                        headers: [
-                            "Content-Type": "application/rss+xml; charset=utf-8"
-                        ],
-                        body: Self.validRSSFeedXML(
-                            channelTitle: "Reappearing Feed",
-                            channelLink: "https://example.com/reappearing/",
-                            language: "en",
-                            itemTitle: "Revived Article",
-                            itemLink: "https://example.com/reappearing/articles/revived",
-                            itemGUID: "revived-article",
-                            itemDescription: "Revived article summary",
-                            pubDate: "Tue, 02 Jan 2024 10:00:00 GMT"
-                        )
-                    )
-                ]
-            )
-        )
-
-        let feed = try #require(try harness.insertFeeds(urls: [feedURL]).first)
-        let refreshedEntryExternalID = ArticleIdentityService.makeExternalID(
-            from: ArticleIdentityInput(
-                feedURL: feedURL,
-                guid: "revived-article",
-                articleURL: "https://example.com/reappearing/articles/revived",
-                title: "Revived Article",
-                publishedAt: FeedDateParsingService.parse("Tue, 02 Jan 2024 10:00:00 GMT")
-            )
-        )
-        _ = try harness.insertArticle(
-            feed: feed,
-            externalID: refreshedEntryExternalID,
-            guid: "revived-article",
-            url: "https://example.com/reappearing/articles/revived",
-            title: "Stale Revived Article",
-            isDeletedAtSource: true
-        )
-
-        let result = await harness.service.refresh(feedID: feed.id)
-
-        #expect(result.status == .fetched)
-        #expect(result.upsertedEntryCount == 1)
-
-        let articles = try harness.articleRepository.fetchArticles(feedID: feed.id)
-        #expect(articles.count == 1)
-
-        let revivedArticle = try #require(articles.first)
-        #expect(revivedArticle.externalID == refreshedEntryExternalID)
-        #expect(revivedArticle.isDeletedAtSource == false)
-        #expect(revivedArticle.title == "Revived Article")
-    }
-
-    @Test
-    func articleStateReadUnreadTransitionUpdatesStateAndInteractionTimestamps() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let feed = try #require(try harness.insertFeeds(urls: ["https://example.com/state-read.xml"]).first)
-        let article = try harness.insertArticle(
-            feed: feed,
-            externalID: "state-read-article",
-            url: "https://example.com/state-read/articles/1",
-            title: "State Read Article"
-        )
-        let baseTime = Date().addingTimeInterval(60)
-        let readAt = baseTime
-        let unreadAt = baseTime.addingTimeInterval(600)
-
-        let readSnapshot = try harness.articleStateService.markAsRead(article: article, at: readAt)
-        let unreadSnapshot = try harness.articleStateService.markAsUnread(article: article, at: unreadAt)
-
-        #expect(readSnapshot.isRead == true)
-        #expect(readSnapshot.readAt == readAt)
-        #expect(readSnapshot.lastInteractionAt == readAt)
-        #expect(readSnapshot.updatedAt == readAt)
-
-        #expect(unreadSnapshot.isRead == false)
-        #expect(unreadSnapshot.readAt == nil)
-        #expect(unreadSnapshot.lastInteractionAt == unreadAt)
-        #expect(unreadSnapshot.updatedAt == unreadAt)
-
-        let persistedSnapshot = try harness.articleStateService.fetchStateSnapshot(
-            feedID: feed.id,
-            articleExternalID: article.externalID
-        )
-        let state = try #require(persistedSnapshot)
-        #expect(state.isRead == false)
-        #expect(state.readAt == nil)
-        #expect(state.lastInteractionAt == unreadAt)
-        #expect(state.updatedAt == unreadAt)
-    }
-
-    @Test
-    func articleStateToggleStarredTransitionsOnAndOffWithFreshTimestamps() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let feed = try #require(try harness.insertFeeds(urls: ["https://example.com/state-star.xml"]).first)
-        let article = try harness.insertArticle(
-            feed: feed,
-            externalID: "state-star-article",
-            url: "https://example.com/state-star/articles/1",
-            title: "State Star Article"
-        )
-        let baseTime = Date().addingTimeInterval(60)
-        let starredAt = baseTime
-        let unstarredAt = baseTime.addingTimeInterval(300)
-
-        let starredSnapshot = try harness.articleStateService.toggleStarred(article: article, at: starredAt)
-        let unstarredSnapshot = try harness.articleStateService.toggleStarred(article: article, at: unstarredAt)
-
-        #expect(starredSnapshot.isStarred == true)
-        #expect(starredSnapshot.starredAt == starredAt)
-        #expect(starredSnapshot.lastInteractionAt == starredAt)
-        #expect(starredSnapshot.updatedAt == starredAt)
-
-        #expect(unstarredSnapshot.isStarred == false)
-        #expect(unstarredSnapshot.starredAt == nil)
-        #expect(unstarredSnapshot.lastInteractionAt == unstarredAt)
-        #expect(unstarredSnapshot.updatedAt == unstarredAt)
-    }
-
-    @Test
-    func articleStateBulkMarkAllVisibleAsReadMarksArticlesAcrossFeeds() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let feeds = try harness.insertFeeds(
-            urls: [
-                "https://example.com/state-bulk-1.xml",
-                "https://example.com/state-bulk-2.xml"
-            ]
-        )
-        let firstArticle = try harness.insertArticle(
-            feed: feeds[0],
-            externalID: "bulk-article-1",
-            url: "https://example.com/state-bulk-1/articles/1",
-            title: "Bulk Article One"
-        )
-        let secondArticle = try harness.insertArticle(
-            feed: feeds[0],
-            externalID: "bulk-article-2",
-            url: "https://example.com/state-bulk-1/articles/2",
-            title: "Bulk Article Two"
-        )
-        let thirdArticle = try harness.insertArticle(
-            feed: feeds[1],
-            externalID: "bulk-article-3",
-            url: "https://example.com/state-bulk-2/articles/1",
-            title: "Bulk Article Three"
-        )
-        let actionAt = Date().addingTimeInterval(60)
-
-        let snapshots = try harness.articleStateService.markAllVisibleAsRead(
-            [firstArticle, secondArticle, thirdArticle],
-            at: actionAt
-        )
-
-        #expect(snapshots.count == 3)
-        #expect(snapshots.allSatisfy { $0.isRead })
-        #expect(snapshots.allSatisfy { $0.readAt == actionAt })
-        #expect(snapshots.allSatisfy { $0.lastInteractionAt == actionAt })
-        #expect(snapshots.allSatisfy { $0.updatedAt == actionAt })
-
-        let persistedSnapshots = try harness.articleStateService.fetchStateSnapshots(
-            for: [firstArticle, secondArticle, thirdArticle]
-        )
-        #expect(persistedSnapshots.count == 3)
-        #expect(Array(persistedSnapshots.values).allSatisfy { $0.isRead })
-        #expect(persistedSnapshots.values.allSatisfy { $0.readAt == actionAt })
-    }
-
-    @Test
-    func articleStateRejectsStaleTransitionWhenUpdatedAtIsOlderThanCurrentState() throws {
-        let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
-        let feed = try #require(try harness.insertFeeds(urls: ["https://example.com/state-lww.xml"]).first)
-        let article = try harness.insertArticle(
-            feed: feed,
-            externalID: "state-lww-article",
-            url: "https://example.com/state-lww/articles/1",
-            title: "State LWW Article"
-        )
-        let newerTimestamp = Date().addingTimeInterval(120)
-        let olderTimestamp = Date().addingTimeInterval(60)
-
-        _ = try harness.articleStateService.markAsRead(article: article, at: newerTimestamp)
-        _ = try harness.articleStateRepository.upsert(
-            feedID: feed.id,
-            articleExternalID: article.externalID,
-            update: ArticleStateUpsert(
-                isRead: false,
-                readAt: nil,
-                lastInteractionAt: olderTimestamp,
-                updatedAt: olderTimestamp
-            )
-        )
-
-        let persistedSnapshot = try harness.articleStateService.fetchStateSnapshot(
-            feedID: feed.id,
-            articleExternalID: article.externalID
-        )
-        let state = try #require(persistedSnapshot)
-        #expect(state.isRead == true)
-        #expect(state.readAt == newerTimestamp)
-        #expect(state.lastInteractionAt == newerTimestamp)
-        #expect(state.updatedAt == newerTimestamp)
-    }
-
-    @Test
     func sourcesSidebarQuerySnapshotAggregatesUnreadAndStarredStateForFeeds() throws {
         let harness = try TestHarness.make(httpClient: ScriptedHTTPClient())
         let feeds = try harness.insertFeeds(
